requirements.txt
googletrans==4.0.0-rc1
pandas==1.3.4
nltk==3.6.5
asyncpg==0.29.0
python-dotenv==0.20.0
SQLAlchemy==1.4.27
Flask==2.1.1
gunicorn==20.1.0
Telethon==1.31.1
aiohttp==3.8.4
fastapi==0.95.1
websockets==11.0.2